Renderer: Refactor project
