+++ WolvenKit.Render/WolvenKit.Render.csproj
@@ -105,6 +105,8 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Animations.cs" />
+    <Compile Include="CommonData.cs" />
     <Compile Include="CSkeleton.cs" />
     <Compile Include="frmRender.cs">
       <SubType>Form</SubType>
@@ -112,9 +114,11 @@
     <Compile Include="frmRender.Designer.cs">
       <DependentUpon>frmRender.cs</DependentUpon>
     </Compile>
+    <Compile Include="Mesh.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ReadMeshData.cs" />
+    <Compile Include="MeshData.cs" />
+    <Compile Include="Rig.cs" />
     <EmbeddedResource Include="frmRender.resx">
       <DependentUpon>frmRender.cs</DependentUpon>
     </EmbeddedResource>
@@ -153,6 +157,10 @@
     <ProjectReference Include="..\WolvenKit.CR2W\WolvenKit.CR2W.csproj">
       <Project>{f3129e02-31ad-4327-a9ef-7c52d777c71f}</Project>
       <Name>WolvenKit.CR2W</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\WolvenKit.W3Strings\WolvenKit.W3Strings.csproj">
+      <Project>{32d40d10-d6c3-48f0-aeb6-0f41c7e75cfd}</Project>
+      <Name>WolvenKit.W3Strings</Name>
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
